+++ source/renderer/app/components/wallet/summary/WalletSummaryAssets.js
@@ -7,9 +7,9 @@ import classNames from 'classnames';
 import BorderedBox from '../../widgets/BorderedBox';
 import styles from './WalletSummaryAssets.scss';
 import Wallet from '../../../domains/Wallet';
-import AssetToken from '../../assets/AssetToken';
+import Asset from '../../assets/Asset';
 import AssetAmount from '../../assets/AssetAmount';
-import type { WalletSummaryAsset } from '../../../api/assets/types';
+import type { AssetToken } from '../../../api/assets/types';
 import LoadingSpinner from '../../widgets/LoadingSpinner';
 
 const messages = defineMessages({
@@ -32,7 +32,7 @@ const messages = defineMessages({
 
 type Props = {
   wallet: Wallet,
-  assets: Array<WalletSummaryAsset>,
+  assets: Array<AssetToken>,
   onOpenAssetSend: Function,
   onCopyAssetItem: Function,
   onAssetSettings: Function,
@@ -88,15 +88,15 @@ export default class WalletSummaryAssets extends Component<Props, State> {
           </div>
         ) : (
           <div className={styles.component}>
-            {assets.map((asset: WalletSummaryAsset, index: number) => (
+            {assets.map((asset: AssetToken, index: number) => (
               <BorderedBox
                 className={styles.assetsContainer}
-                key={asset.policyId + asset.assetName + asset.fingerprint}
+                key={asset.uniqueId}
                 onMouseEnter={this.handleHoverAsset}
               >
-                {asset.fingerprint && (
+                {asset.uniqueId && (
                   <div className={styles.assetsLeftContainer}>
-                    <AssetToken
+                    <Asset
                       asset={asset}
                       onCopyAssetItem={onCopyAssetItem}
                       metadataNameChars={get('name', asset.metadata, 0)}
@@ -117,7 +117,7 @@ export default class WalletSummaryAssets extends Component<Props, State> {
                     </div>
                   </div>
                 )}
-                {asset.fingerprint && (
+                {asset.uniqueId && (
                   <div className={styles.assetRightContainer}>
                     <button
                       className={classNames([